daemon: make drain behaviour modular
Refactored drain logic and moved drain related
functions into drain.go for easier maintenance.
